Fixes #2: supervisord logs and socket moved out from /data volume
This resolves the conflict when /data volume is used and it is used by more then one running container.

The supervisord logs are available via 'docker logs' and that is probably best way to source them from outside the container. If systemd/journald is available in the system, it's even easier to access them via these tools.